(Feature) [xDai] - Contract interaction abi lookup (#1429)
