--- src/logic/contractInteraction/sources/index.ts
@@ -1,7 +1,0 @@
-import EtherscanService from 'src/logic/contractInteraction/sources/EtherscanService'
-
-const sources = {
-  etherscan: new EtherscanService({ rps: 4 }),
-}
-
-export const getConfiguredSource = () => sources['etherscan']
